requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.1
google-generativeai==0.5.2
numpy==1.26.4
pandas==2.2.2
nltk==3.8.1
